Build 1.0.0.59
Fix: Half-Orc: Improved Power Attack now correctly costs 1ap per rank (Reported by Question2)
Fix: Making the number of available build points negative (e.g. full spend 32 point build then switched to 28 point) no

longer allows you to spend additional build points
Fix: Minor graphical error in passive enhancement available icon fixed
Fix: Self and Party Buff "Inspire Excellence" no longer erroneously grants +5 Morale bonus to constitution and a -2

Penalty to AC (Reported by Kamdragon)
A general report of issues with the Kensei tree was reported by Question2. The following fixes were done after a full

review of the tree.
---Fix: Kensei: Kensei Focus - selector descriptions updated
---Fix: Kensei: Spiritual Bond - description updated
---Fix: Kensei: Strike With No Thought - description updated
---Fix: Kensei: Power Surge - description updated
---Fix: Kensei: One Cut - description updated
---Fix: Kensei: Alacrity - description updated
---Fix: Kensei: Weapon Group Specialization I - selector descriptions updated - effects updated
---Fix: Kensei: Weapon Group Specialization II - selector descriptions updated - effects updated
---Fix: Kensei: Ascetic Training - descriptions updated
---Fix: Kensei: Weapon Group Specialization III - selector descriptions updated - effects updated
---Fix: Kensei: Ability I/II - description updated
---Fix: Kensei: Weapon Group Specialization IV - selector descriptions updated - effects updated
New: Feat: Suggestion (Bard feat) now has a DC entry
Arcanotechnician: All tier 1 enhancements now have 1 Artificer level as a requirement
Battle Engineer: All tier 1 enhancements now have 1 Artificer level as a requirement
Renegade Mastermaker: All tier 1 enhancements now have 1 Artificer level as a requirement
Frenzied Berserker: All tier 1 enhancements now have 1 Barbarian level as a requirement
Occult Slayer: All tier 1 enhancements now have 1 Barbarian level as a requirement
Ravager: All tier 1 enhancements now have 1 Barbarian level as a requirement
Divine Disciple: All tier 1 enhancements now have 1 Cleric level as a requirement
Ranger: Arcane Archer missing 1 Ranger level requirement added

U41 Changes:
---Old Eldritch Knight tree removed
------New "Eldritch Knight (Wiz)" tree added and renamed from (Lamannia) as now live
------New "Eldritch Knight (Sorc)" tree added and renamed from (Lamannia) as now live
---Spell descriptions updated for:
------Acid Spray
------Burning Hands
------Shocking Grasp
------Scorch (Also no longer has incorrect spelling)
------Snowball Swarm
------Acid Blast
------Fireball
------Lightning Bolt
------Fire Trap
------Negative Energy Burst
------Ball Lightning
------Cone of Cold
------Chain Lightning
------Otiluke's Freezing Sphere
------Necrotic Ray
------Delayed Blast Fireball
------Polar Ray
------Meteor Swarm
---Prowess' 5 Piece Set is now: Whenever you activate an Action Boost, gain +50 Melee Power for 10 seconds.
---The Enlightened Step's 5 Piece Set is now: +5% Doublestrike and Doubleshot. (Missing effects added)
---Feat: Arcane Warrior duration text changed from 6 to 12 seconds. Spell power bonus changed to +2 (from 1)
---Remnant turn in tomes now go up to +4
------Inherent Magical Resistance
------Inherent Melee Power
------Inherent Physical Resistance
------Inherent Ranged Power
------Inherent Spell Power
---Soulrazor now has DR Crystal listed as being bypassed
Bards:
---Fatesinger Tree updates:
------Fatesinger: Allure - Description and effects updated
------Fatesinger: Tailwind - Description updated
------Fatesinger: Lucidity - Description and effects updated
------Fatesinger's Repertoire - now grants "Bardic Ballad" instead of "Inspire Courage"
---New Automatic Bard Feat "Bardic Ballad" added, automatically acquired at Bard level 1
---New Automatic Bard Feat "Bardic Inspiration" added, automatically acquired at Bard level 3
---New Automatic Bard Feat "Soothing Song" added, automatically acquired at Bard level 8
---Bard feat "Inspire Greatness" reworked into "Ballad Melody: Greatness", automatically acquired at Bard level 9
---Bard feat "Song of Freedom "- minor text update
---New Trainable Bard Feat "Improved Bardic Music" added, Min level 12 Bard, 15+ Perform
---Bard feat "Inspire Heroics" updated to act passively
---Old Bard Feats Retired:
------Inspire Competence
------Inspire Courage
---Bard Self and Party buffs updated, renamed and new ones added:
------Bard: Inspire Courage
------Bard: Arcane Aid
------Bard: Inspire Competence
------Bard: Inspire Greatness
------Bard: Inspire Heroics
------Bard: Arcane Shield Chant
------Bard: Ironskin Chant
------Bard: Reckless Chant
------Bard: Expeditious Chant
------Bard: Chant of Power
------Bard: Inspire Excellence
------Bard: Spell Song Trance
------Bard: Song of Arcane Might
---Bard Warchanter updates:
------Warchanter: Skaldic rage now correctly requires 1 Bard level
------Warchanter: All tier 1 enhancements now have 1 Bard level as a requirement
------Warchanter: Skaldic rage - description updated
------Warchanter: Song of Heroism - description updated and effects added
------Warchanter: Fighting Spirit - Description and effects updated
------Warchanter: Victory Song - Description updated
------Warchanter: Warmaster - description updated and effects/dc added
------Warchanter: Inspired Bravery - effects added
------Warchanter: The Poetic Edda - description updated and effects added
------Warchanter: Rough and Ready - description updated and effects added
------Warchanter: Focused - description updated and effects added
------Warchanter: Words of Encouragement - description updated and effects added
------Warchanter: Ballad Melody: Recklessness - description updated and effects added
------Warchanter: Northwind - description updated
---Bard: Spellsinger updates:
------Spellsinger - description and effects updated
------Spellsinger now correctly requires 1 Bard level
------Spellsinger: All tier 1 enhancements now have 1 Bard level as a requirement
------Spellsinger: Music of the Sewers - description updated
------Spellsinger: Music of the Dead - description updated and effects added
------Spellsinger: Music of the Makers - description updated and effects added
------Spellsinger: Virtuoso - description updated and effects added
------Spellsinger: Maestro of Life and Death - description updated and effects added
------Spellsinger: Lingering Songs - description updated and effects added
------Spellsinger: Reviving Verse - description updated and effects added
------Spellsinger: Sustaining Song - description updated and effects added
------Spellsinger: Frolic - description updated
------Spellsinger: Prodigy - description updated and effects added
------Spellsinger: Spell Song Vigor - description updated and effects added
---Bard: Swashbuckler updates:
------Swashbuckler: Confidence now correctly requires 1 Bard level
------Swashbuckler: All tier 1 enhancements now have 1 Bard level as a requirement
------Swashbuckler: Swashbuckling - description updated and effects added

+++ DDOCP/AbilitySpend.cpp
@@ -15,7 +15,7 @@ namespace
     const unsigned f_verCurrent = 1;
 
     size_t c_maxSpend = 10;
-    size_t c_statIncreaseCosts[] = {1, 1, 1, 1, 1, 1, 2, 2, 3, 3};
+    int c_statIncreaseCosts[] = {1, 1, 1, 1, 1, 1, 2, 2, 3, 3};
     const unsigned arraySize = sizeof(c_statIncreaseCosts) / sizeof(*c_statIncreaseCosts);
     BOOST_STATIC_ASSERT(arraySize == 10);
 }
@@ -103,8 +103,8 @@ bool AbilitySpend::CanSpendOnAbility(
     {
         // its not at max, but are there enough points left to
         // spend to the next level?
-        size_t availablePoints = (m_AvailableSpend - PointsSpent());
-        size_t nextCost = c_statIncreaseCosts[currentSpend];
+        int availablePoints = ((int)m_AvailableSpend - (int)PointsSpent());
+        int nextCost = c_statIncreaseCosts[currentSpend];
         canSpend = (availablePoints >= nextCost); // true if enough points available
     }
     return canSpend;
